split NavSchema
